[Decode] fix some kw issues (#1322)
fix some dec kw issues

Issue: -
Test: manual

+++ _studio/shared/umc/codec/vc1_common/src/umc_vc1_common.cpp
@@ -48,7 +48,7 @@ namespace UMC
                 if (0 == counter)
                     iCur = src[i];
                 iCur <<= 8;
-                iCur |= src[i];
+                iCur |= (uint32_t)src[i];
                 ++counter;
             }
         }
